try to refactor codes for create new fragment TaskContainFragment
