+++ sharedbaseclass/src/main/res/layout/shared_tasks_container_view.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+              android:layout_width="match_parent"
+              android:layout_height="match_parent"
+              android:orientation="vertical">
+
+  <include layout="@layout/shared_header_view"></include>
+
+  <android.support.v7.widget.RecyclerView
+    android:id="@+id/list"
+    android:layout_width="match_parent"
+    android:layout_height="0dp"
+    android:layout_weight="1"
+    android:background="@android:color/white"
+    ></android.support.v7.widget.RecyclerView>
+  <TextView
+    android:id="@+id/no_task_page"
+    android:layout_width="match_parent"
+    android:layout_height="0dp"
+    android:layout_weight="1"
+    android:gravity="center_vertical|center_horizontal"
+    android:background="@android:color/white"
+    android:visibility="gone"
+    android:text="@string/fragment_tasks_container_there_is_no_task_here"
+    />
+
+</LinearLayout>
